pyproject.toml
[tool.poetry]
name = "airflow_dvc"
version = "1.9.8"
description = "DVC operator for Airflow"
authors = ["Piotr Styczyński <piotrsty1@gmail.com>"]
readme = "README.md"
include = [
    { path = 'README.md', format = 'wheel' },
    { path = 'README.md', format = 'sdist' },
    { path = 'example/dags/*', format = 'wheel' },
    { path = 'example/dags/*', format = 'sdist' },
    { path = 'example/plugins/*', format = 'wheel' },
    { path = 'example/plugins/*', format = 'sdist' },
    { path = 'pyproject.toml', format = 'wheel' },
    { path = 'pyproject.toml', format = 'sdist' },
    "README.md",
    "example",
    "example/dags",
    "example/dags/dvc_sensor_example.py",
    "example/dags/dvc_upload_example.py",
    "example/plugins",
    "example/plugins/dvc.py",
    "pyproject.toml"
]

[tool.poetry.scripts]
publish = 'publish:publish'
airflow_dvc = 'airflow_dvc.cli.entrypoint:run_cli'

[[tool.poetry.source]]
name = "default_pypi"
url = "https://pypi.python.org/simple/"
secondary = true

[[tool.poetry.source]]
name = "covid_genomics_pypi"
url = "https://pypi.covidgenomics.com/production/"
default = true

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "<1.4.0"
GitPython = "^3.1.14"
apache-airflow = ">=2.0.0"
apache-airflow-providers-amazon = "^1.3.0"
typer = "^0.3.2"
semver = "^2.13.0"
semantic-version = "^2.8.5"
toml = "^0.10.2"
dvc-fs = "^0.7.2"
dvc = {extras = ["s3"], version = "^2.5.4"}
mmseq-api-client = "^1.2.0"

[tool.poetry.extras]
dvc = ["s3", "dvc>=2.0.18"]

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-xdist = "^2.2.0"
isort = "^5.7.0"
black = "^20.8b1"
pydoc-markdown = "^3.9.0"
s3pypi = "^0.11.0"
flakehell = "0.9.0"
flake8 = "3.8.3"
poetry-publish = "^0.4.1"
bump2version = "^1.0.1"
gitchangelog = "^3.0.4"

[tool.poetry.plugins]

[tool.poetry.plugins."airflow.plugins"]
"airflow_dvc_plugin" = "airflow_dvc:DVCPlugin"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flakehell]
exclude = ["README.rst", "README.md"]
format = "colored"
max_line_length = 120
show_source = true
whitelist = "../../allowlist.txt"

[tool.flakehell.plugins]
flake8-bandit = ["+*", "-S322"]
flake8-bugbear = ["+*"]
flake8-builtins = ["+*"]
flake8-comprehensions = ["+*"]
flake8-darglint = ["+*"]
flake8-docstrings = ["+*"]
flake8-eradicate = ["+*"]
flake8-isort = ["+*"]
flake8-mutable = ["+*"]
flake8-pytest-style = ["+*"]
flake8-spellcheck = ["+*"]
mccabe = ["+*"]
pep8-naming = ["+*"]
pycodestyle = ["+*", "-E203", "-W503"]
pyflakes = ["+*", "-E203"]
pylint = ["+*", "-E203"]

[build-system]
requires = ["poetry-core>=1.1.2a3"]
build-backend = "poetry.core.masonry.api"
